moved dsl to the resources namespace
