+++ lib/json_api.rb
@@ -1,7 +1,6 @@
 module JSONAPI
 end
 
-require_relative 'json_api/dsl'
 require_relative 'json_api/resource'
 require_relative 'json_api/exceptions'
 require_relative 'json_api/configuration'
